Order Service basics apis ready

+++ customer-service/src/main/java/com/easyelectroshop/customerservice/DTO/Customer/Customer.java
@@ -16,6 +16,6 @@ public record Customer(
         boolean blocked,
         boolean active,
         int totalOrders,
-        int totalOrderAmount
+        double totalOrderAmount
 ) {
 }
